+++ NEWS.md
@@ -11,6 +11,7 @@
 
 ### Bug fixes
 * Remove expired data cleanup from tenant update ([MODMARCMIG-104](https://folio-org.atlassian.net/browse/MODMARCMIG-104))
+* Implement spring batch partitioning to support parallel processing in Spring Batch 6 ([MODMARCMIG-107](https://folio-org.atlassian.net/browse/MODMARCMIG-107))
 
 ### Tech Dept
 * Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))